Get docker working (#1)
Yesterday we just focused on getting notes-sdk working, building with
that. This PR integrates what we built into the rest of the SAF
framework template.

The upshot of these changes are that you can:
* Go to `deploy/dev`, run `npm run dev -- --build` and you'll get the
full stack running in dev mode
* Go to `deploy/prod`, run `npm run build && npm run prod-local` and
you'll get the whole stack running in "prod" mode... locally.

The main diff between these two is "dev" runs the vite server, and
"prod" builds the static assets with vite then has Caddy serve those
assets. Also, "prod-local" does not keep state between runs for the sake
of playwright testing, but that only affects the identity server user
accounts currently.

The full stack includes:
* Caddy
* The generated "notes" service, even though we're not using it right
now, we'll want to add the web socket server to it anyway.
* Vite if in dev mode
* Prometheus
* Grafana
* node_exporter
* Loki (if in prod)
* Identity server

If you run either dev or prod, you can:
* Go to [http://docker.localhost](http://docker.localhost) to see the
root-app's homepage
* Go to [http://auth.docker.localhost](http://auth.docker.localhost) to
log in
* Go to [http://app.docker.localhost](http://app.docker.localhost) to
see what we've built
* Go to
[http://notes.docker.localhost/health](http://notes.docker.localhost/health)
to check on the health of the express app
* Go to
[http://grafana.docker.localhost/](http://grafana.docker.localhost/) and
look around. Could also add the prometheus and/or loki sources, though I
haven't tried that.

ALSO:
* Remove "@your-org/" from package names
* Refactor SDK a bit to export the key bits to the SPA. At some point
should probably export the pages instead of the router, or even just
components, but not necessary right now.
* Fix playwright.